eslint-changed: Improve eslint 9 support
eslint-changed was only using eslintrc for v8 and flat config for v9.
Instead it should detect which to use in the same way eslint does.

To make use of eslint 9's experimental config file resolution, you need
to pass `--flag unstable_config_lookup_from_file` to eslint. In the API,
the equivalen is passing an object with
`flags: [ "unstable_config_lookup_from_file" ]` to the ESLint class's
constructor. To make this possible with eslint-changed, we add a
`--eslint-options` option.

Then, in preparation for eslint 9, let's update the tests to use flat
config (plus one testing that eslintrc still works).