eslint: Clean up deprecated rules
There are a lot of rules in our config that are normally disabled by
eslint-plugin-prettier/recommended, but we've been turning them back on.
Many of them are now deprecated in eslint too (in favor of
`@stylistic/eslint-plugin-js`). Let's disable them and let prettier
handle it.

- comma-spacing
- computed-property-spacing
- curly
- func-call-spacing
- jsx-quotes
- key-spacing
- keyword-spacing
- no-extra-semi
- no-multiple-empty-lines
- no-multi-spaces
- no-spaced-func
- no-trailing-spaces
- object-curly-spacing
- operator-linebreak
- padded-blocks
- react/jsx-curly-spacing
- react/jsx-tag-spacing
- semi
- semi-spacing
- space-before-blocks
- space-infix-ops
- space-in-parens
- space-unary-ops

A few more we had been disabling even though prettier already disabled
them.

- comma-dangle
- space-before-function-paren

And then there are a few disabled in favor of `eslint-plugin-n` instead.

- no-process-exit → n/no-process-exit
- no-restricted-imports → n/no-restricted-import
- no-restricted-modules → n/no-restricted-require

And since we're now bringing in `eslint-plugin-n`, may as well enable a
few additional rules from that plugin too.

- n/no-deprecated-api
- n/no-exports-assign
- n/no-process-exit
- n/no-restricted-import
- n/no-restricted-require
- n/process-exit-as-throw