ci: Create pr-update-to-project tag for new projects
Previously this happened because the creation of the new project's
CHANGELOG.md was seen as touching the changelog, but we removed
"changelog touched" updating the tag in #38974.

To ensure the tag gets created when the project is (because other stuff
expects it), look for the addition of a composer.json file.