js-packages: Fix some JS lints ahead of the eslint 9 upgrade
To make the eslint 9 upgrade PR smaller, let's fix some of the lints
that it'll complain about ahead of time.

Of particular note is that, when we added 'plugin:@wordpress/jsx-a11y'
in #38436, we accidentally left 'plugin:jsx-a11y/recommended' in place
after it which overrides some of the configuration that
'plugin:@wordpress/jsx-a11y' does. Fixing that disables a few rules we
had been inline-ignoring in various places: jsx-a11y/media-has-caption,
jsx-a11y/no-noninteractive-tabindex, and
jsx-a11y/role-has-required-aria-props.