Publicize: Use get_all_connections_for_user in wpcom/v2 endpoints
Expanding on #40589 this attempts to refactor the wpcom/v2 endpoints and
bring the get_all_connections_for_user method to publicize-base.

It means we start relying on the XMLRPC calls entirely, which could be a
performance problem with testing the connections.

There's probably some more refactoring we could do to proxy the
endpoints and use those within Publicize, by overriding some of the
methods, like get_publicize_conns_test_results