Add a "wordpress-dataviews-snapshot" package
This package contains a snapshot build of `@wordpress/dataviews/wp` and
metadata needed to have `@wordpress/dependency-extraction-webpack-plugin`
extract the dependency, all in a manner that should work even if
multiple plugins include different versions of the package.

This also updates the monorepo config to encourage use of this package
instead of attempting to use `@wordpress/dataviews` directly.