ci: Check for use of SCSS `random()`
Normal CSS can't do random numbers. SCSS has a `random()` function,
which produces random values in the output CSS. Unfortunately this makes
the built CSS change for each build, which means that every PR merged
will require a Simple deploy. To avoid this, add a CI check looking for
calls to the function in all `.scss` files.

Also do the same for `unique-id()`.