Priority indicators: update logic to match new priority matrix
We now ask different questions in our issue templates:
https://github.com/Automattic/jetpack/pull/40501

This commit updates the logic to match the new questions and answers. It also captures the optional extra information ("other impact(s)") an issue can have, since that info can bump the severity, and in turn the priority, of an issue.

See pfVjQF-su-p2