Stop quick links disappearing after using Quick Edit
The post list has a number of quick links including 'Promote with
Blaze', 'Copy' and Quick Edit. After using Quick Edit the Blaze and Copy
links go missing - they're not registered when the ajax response
includes the new quick links.

This is solved by registering the `post_row_actions` hook on the
admin-ajax.php page as well as the edit.php page.