Blocks: display custom messages for blocks when module is disabled
Since we introduced the Blocks module in #39449, some folks may find themselves with Jetpack active on their site, and they may have added blocks to their posts or site templates before to turn on the module.
Alternatively, they may use a theme that comes with Jetpack blocks in some of its templates, but Jetpack's Offline mode may be active on the site.

In those scenarios, let's customize the error displayed instead of the block.