Clean up names in package.json files
We have a bunch of package.json files that are never meant to be
published to npmjs.com. Many are for JS dependencies to build and bundle
JS in our GitHub Actions, PHP packages, and WordPress plugins. Others
are for tooling that's used within the monorepo.

Every once in a while someone notices one of these where we have an
unscoped "name", either unused or for internal use, and tries to submit
a bug bounty for it despite the intra-monorepo `workspace:*` deps not
allowing that.

To avoid that annoyance, this adds a check in `lint-project-structure.sh`
that every package.json "name" must either be scoped to `@automattic/*`,
or begin with an underscore (which is not allowed to be published to
npmjs.com), or be an existing package that's maintained by either us
(specifically, our `matticbot` account) or NPM itself (`npm`).

Then we also have to fix the existing instances someone hasn't caught
yet:
* For most we just delete the unused "name".
* For `jetpack-cli` in `tools/cli/package.json`, we manually published a
  dummy package similar to existing ones like for `jetpack-js-tools`.
* For `jetpack-e2e-commons` in `tools/e2e-commons/package.json`, we
  renamed that to `_jetpack-e2e-commons`. Which unfortunately required a
  lot of references to be updated, sigh.