Display Blaze quick action after editing pages/products
After using quick edit to save a change the Blaze quick action was no
longer displayed. It works on posts but not on other CPTs, including
pages.

This change allows the Share with Blaze on all blaze-eligible CPTs
after a quick-edit.

Note that Blaze is only designed for posts, pages and products at
present, nothing in this PR changes that.

Related: https://github.com/Automattic/wp-calypso/issues/97995