Social: Enable the new admin page for the Social plugin.
Currently we have a feature check that is enabling the feature if the
Social plugin is active. That is causing some problems where the admin
page doesn't show up until the feature information is updated for the
site, after the site is connected etc.)

This change sets the constant in the Social plugin, so the feature is
always enabled from now on with the Social plugin active. We can then
remove the version check from the feature check in WPCOM.