tools: Force releases of js-package dependents
Our js-packages are usually bundled when included in (php-)packages and
plugins, rather than being used directly at runtime. This can lead to an
edge case in releases, if a js-package has been updated and the plugin
being released depends on that, but the php-package that bundles the
js-package has no changes and so does not get released.

To avoid this, when a js-package is being prepared for release, any of
its dependents should also be released. We go about that using a
solution similar to that from #33665, which fixed a similar edge case.