ci: Add support for "don't care about coverage" label
One feature we want for the code coverage check is the ability to easily
(but intentionally) override the check for PRs where getting things done
is more important than code coverage.

We've decided to do this by allowing for a label, currently named
"I don't care about code coverage for this PR", to be added to the PR to
override the check.