cli: `jetpack rsync`'s untracked files check should check packages too
PR #41442 added messaging for when `jetpack rsync` is going to ignore a
file because it hasn't been `git add`ed. But that PR only checked files
directly in the plugin being synced, which is not terribly useful for
mu-wpcom-plugin where everything happens in packages/jetpack-mu-wpcom.

To fix this, for the check, look at any monorepo packages symlinked from
`vendor/*/*` and `jetpack_vendor/*/*` too. Further recursion is not
necessary, since Composer already hoists everything relevant.