Use Verbum for editing comments on simple sites
Verbum was used in Calypso for editing comments, this change brings it
to wp-admin on the edit comment screen too. People managing comments
should have a nice wyiswyg way of editing comments written using Verbum
just like the commentators did. The alternative is for them to manually
edit Gutenberg mark-up

This change does not bring Verbum to comment quick-edit at this time.