ci: Test for wpcom Simple filename restrictions
Wpcom Simple has several restrictions on files that can be committed. To
help avoid problems during the deploy, let's add a CI run reproducing
these checks against the relevant plugin builds.

* Underscores are not allowed in new filenames (with some limited
  exceptions).
* New file and directory names can only use a very limited set of
  characters.
* New and modified files must not be symlinks.

There's also a check for executable files in the code, but that one
doesn't seem to be enabled in the wpcom repo (probably because it'd
break a bunch of stuff in `bin/`).