Use load_theme_textdomain to load translations from WP_LANG_DIR/mu-plugins
During testing, I attempted to use load_muplugin_textdomain to load translations for
jetpack-mu-wpcom with the following approach:

    load_muplugin_textdomain( 'jetpack-mu-wpcom', '../languages/mu-plugins' );

However, this method relied on a relative path resolving to WP_LANG_DIR/mu-plugins.

To improve reliability, we switched to load_theme_textdomain with an absolute path:

    load_theme_textdomain( 'jetpack-mu-wpcom', WP_LANG_DIR . '/mu-plugins' );

The path corresponds to the download+extraction location of the language packs in the
Jetpack_Mu_Wpcom::maybe_update_translations() method.