phpunit: Static-ify data providers
PHPUnit 11 requires that all `@dataProvider` functions are `static`.
Most of the time this is straightforward, as most of the time data
providers return a static array of test cases.

`WafRuntimeTargetsTest` was a major exception: it's building
`Waf_Runtime` objects for each case, often with mocks, and PHPUnit
doesn't provide methods to create mocks statically. What I wound up
having to do there was return a factory method instead, which would take
`$this` as a parameter to use to create the mocks.

Jetpack also had a few that were somehow passing despite trying to use
`$this->var` before that got set by `set_up()`. And a couple of data
providers taking an unused `$name` parameter that is no longer passed.