phpunit: Add fix trait for WP_UnitTestCase + PHPUnit 11
`WP_UnitTestCase` tries to use PHPUnit's annotation-reading machinery to
look for its `@expectedDeprecated` and `@expectedIncorrectUsage`
annotations: If `$this->getAnnotations()` exists (PHPUnit <9.5) it calls
it, otherwise it tries to use `PHPUnit\Util\Test::parseTestMethodAnnotations()`.

Neither of those exist anymore in PHPUnit 11. But fortunately it prefers
`$this->getAnnotations()` and does it via an existence check, so all we
need to do is provide an implementation of that method via a trait.

Since it's only three of our plugins that use `WP_UnitTestCase`, and we
don't really want to encourage more, I just copy-pasted the trait into
those three plugins instead of setting up a package and so on to share
it.

(Our polyfill here won't work for PHPUnit 12, but we'll deal with that
later. 😅)