Fix search result count
This commit fixes a bug with how search results are displayed. Currently,
combinging hasQuery with hasCorrectedQuery causes the if that was on line
70 to return false, which triggers the default return of the function, around
line 118 with "Showing popular results" (which should just be a default state
on first load, etc.).

This adds "smarter" logic handling so that we show the correct "Founx X result/s"