phpunit: Remove unnecessary uses of `@before`, `@after`, etc
Back when we still supported PHP <=7.1, we had to add these in a lot of
places to work around the fact that PHPUnit 8 started declaring a `void`
return type on `setUp()`, `tearDown()`, and so on, which 7.1 and earler
don't support.

But we've since dropped support for PHP <7.2, so we can swtich back to
the normal functions. It's more straightforward to do that than to start
adding attributes (`#[Before]` and such) for PHPUnit 11+ where the
annotations are deprecated.