Social: Always register the product info endpoint.
With only Jetpack active, the admin page attempts to use the
social-product-info endpoint to retrieve the planning pricing, but it's
not regsitered if the Publicize module isn't active.

This would also be a problem with the new version of the Social plugin,
but the current one registers the endpoints regardless.

This change ensures that we have the endpoint registered, even if the
Publicize module is inactive.