phpunit: Move `@covers` to class level
PHPUnit 11 is implicitly deprecating use of `@covers` at the method
level, in that it deprecated the `@covers` annotation in favor of
attributes like `#[CoversClass()]` and `[#CoversMethod()]` but only
makes the attributes valid at the class level.

Since it seems less likely to break things when not all methods had
`@covers`, and avoids long lists at the top of the class, all
`@covers Class::method` are being changed to just `@covers Class` when
raised to the class level.

Also, since they're not terribly useful anymore without method-level
`@covers`, remove all `@coversDefaultClass`.