Forms: Redirect to /responses when the site has form responses
In the Inbox view, we redirect to the /landing when there are no form
responses. If a response comes in, and the admin refreshes this page,
they don't see it, as they remain on the /landing page.

This change redirects to the responses view if responses are found while
rendering the /landing page.