beta: Fix Jetpack on WoA sites
PR #42965 added a filter in wpcomsh that tries to ensure that
Jetpack-the-plugin (meaning slug `jetpack/jetpack.php`) is always
enabled. This conflicts with the Beta Tester which may deactivate it in
favor of `jetpack-dev/jetpack.php`.

To avoid this, register a hook with a priority just after theirs which
checks if both `jetpack/jetpack.php` and `jetpack-dev/jetpack.php` are
active and deactivates the former if so.