phpunit: Enable reporting and failing on PHPUnit 11 deprecations
Now that we've fixed all the PHPUnit 11 deprecations, adjust the configs
to enable `displayDetailsOnPhpunitDeprecations` and
`failOnPhpunitDeprecation`.

For plugins/crm, plugins/jetpack, and plugins/wpcomsh, there's one more
deprecation left that's also fixed in this PR: WP Core's obsolete
`WP_UnitTestCase_Base::checkRequirements()` function has a doc comment
that PHPUnit is mis-parsing to see a `@group` annotation that's not
actually intended. We can override it to stop PHPUnit from seeing that
doc comment.