Make Jetpack AI work for site editors
The initial approach checked if there was a site connection and made use
of the request_as_blog method. In this iteration, I'm using a different
approach by checking if the site has an admin connected and went back to
use request_as_user method as before. The main reason for the change of
approach is that it seems less risky this way–we still rely on the user
connection but we just require an admin to have the user connection.