Fix types for components package
Ideally, the components package should be standalone and portable, unaware of the logic of other packages, but unfortunately, that is not the case with these utilities. They incorporate the logic from other packages, which we try to fix by wrongly polluting the global TS namespace. Ultimately, we should get rid of these utilities from this package and move to some other package like shared extensions util