Replace __() with dummy param with _x() to make TS happy.
To avoid bad minification, we often add a third dummy param to `__()` but TS doesn't like that because the function accepts only two arguments. So, a better approach is to use `_x()` where you can pass the second argument as the context which has a benefit of giving context for the translators and also makes TS happy.