Make Jetpack AI requests as blog instead of user
In order to allow non-admin users to use Jetpack AI we need to make the
request as blog instead of user because the editor won't have connection
token. Luckily, we don't need the user_id in any of the Jetpack AI
endpoints as we're only relying on the blog_id.