Replace `@wordpress/dataviews` with `@automattic/dataviews`
This is the direction we're going with that package in the near-ish
term.

Then it turns out the old hack for the ESM part of
WordPress/gutenberg#67897 doesn't seem to work anymore, since the new
package doesn't use Babel's `interopRequireDefault` which makes
something blow up trying to access a `.default`. I came across a new
hack that seems to work to convince Node that the `.js` files in the ESM
dirs are ESM (versus the old hack that patched references to the ESM
dirs out of package.json entirely).

But then we have to set `.resolve.fullySpecified` to false for the
package in a bunch of places because the provided ESM build doesn't
include extensions on some of its imports. Unsure whether that would
have existed in the old package too or not.

Also the Protect plugin now needs transpilation of the package added
with the new hack. I guess it's not tree-shaking out as much anymore for
some reason? Which leaves in a bunch of dataviews i18n that has no text
domain set.

The stuff not shaken out apparently includes a "Delete" string
(referring to `.key` on a keyboard event) somewhere. This makes the
i18n-check webpack plugin complain because there's a `__( 'Delete' )`
from js-packages/scan that _is_ still shaken out. So I changed that to
`_x()` with a context to unconfuse that.