phpcs-filter: Fix code coverage run with phpcs 3.13
PHPCS 3.13.0 starts defining some PHP 8.4 token constants if they don't
exist, but defines them as strings. Meanwhile nikic/php-parser complains
if they're not ints.

The problem only occurs for this package with the coverage run, because
it depends on phpcs loading its replacement constants (which only
happens here because of the weird way we set things up to be able to
test the filter) and on PHPUnit generating code coverage.

We can preemptively define the constants before the require to avoid the
problem, so let's do that for now.