masterbar: Resolve sass deprecation warnings
Mostly this is changing `@import` to `@use`. Which turned out to be
somewhat complex here.

The general setup before was that each color scheme had a `colors.scss`
entry point, which did the following:
 1. Import `_variables.scss` to define a bunch of global variables.
 2. Import `../_admin.scss`. That in turn does:
    1. Import `_variables.scss` to process all the global variables,
       setting any that are unset to default values.
    2. Output a bunch of CSS.
    3. Import `_overrides.scss`, `_gutenberg.scss`, and
       `_inline-help.scss` to output a bit more CSS. These mostly use
       different global variables than `_variables.scss` set, with no
       defaulting.
 3. Import `sidebar-notice.scss`, which is also its own entry point, to
    get a bit more CSS.

None of the global variable stuff works with `@use`. What I wound up
doing was changing the `colors.scss` to do like this instead:
 1. Use `_variables.scss`.
 2. Use the `meta.load-css()` mixin to load `../_admin.scss`, passing it
    all the variables from `_variables.scss`. `../_admin.scss` now does:
    1. All the processing and defaulting logic formerly in its
       `_variables.scss` gets moved into `_admin.scss` itself.
    2. We also have to "declare" the variables that `_admin.scss` itself
       doesn't use but its submodules do. And the variables that
       `sidebar-notice.scss` uses, even though `_admin.scss` doesn't use
       them at all.
    3. Outputs the CSS, same as before.
    4. Then it uses `meta.load-css()` to load the other three
       sub-files, passing them the variables that they use.
 3. Use `meta.load-css()` to load `sidebar-notice.scss` too. It handles
    its own variables.

The alternative to the above would have been to turn the guts of
`_admin.scss` and its submodules into mixins. The `colors.scss` files
would `@use` the files and then call the mixins in much the same way
that the files are called now with `meta.load-css()`. The variable
handling might have been a little more straightforward because the
variable map could have been passed through and used more directly. Or
it might have been less straightforward in other ways. 🤷 See
https://sass-lang.com/documentation/at-rules/use/#with-mixins for more
about that method.

And then there was one other deprecation, a use of `color.hue()` needed
changing to `color.channel()`. 😀