search pkg: Fix sass deprecation warnings (#43713)
This was pretty straightforward overall.

We couldn't fix the `@import` of
`@wordpress/block-editor/src/components/colors-gradients/style` in
`src/customberg/lib/_base-styles.scss` though, as that would need to
wait for Gutenberg to convert that file to `@use`. However, as far as I
can tell there's no need for the import at all, as all the same styles
come in from the `@wordpress/block-editor` package itself. Removing the
whole file seems to make no difference to the rendering.