jetpack: Fix Sass deprecation warnings
The changes here are mostly straightforward, although getting there was
somewhat of a pain due to the interconnections of a lot of the different
files. Sass's migrator wanted to overwrite variables in a bunch of
places that didn't really work.

In a few places I went with `@include meta.load-css()` rather than
`@use`. As far as styles go these are equivalent; the former doesn't
make any functions, variables, or mixins available to the current
module, but in these cases there aren't any of those and it lets us
avoid having to `@use ... as style2`, `@use ... as style3`, etc.

There are some deprecation warnings coming from
`@automattic/social-previews` that will need to be fixed upstream.