Inline Search: Address Blaskan theme body class conflict with inline highlighter (#43797)
* Address Blaskan theme fatal

This commit addresses the Blaskan theme fatal by moving away from body_class() based checking, to theme name and theme template checking.
We use this to return the $title early in filter_highlighted_title(), both for the Blasken theme, and p2020, which loads o2 as a plugin and
which calls the_title filter incorrectly (which previously caused its own fatal).

* changelog