Replace renamed `pkg-dir` → `package-directory`
In packages/subscribers-dashboard it seems this isn't needed at all. It
was apparently copy-pasted from packages/jetpack-mu-wpcom. Remove it.

For packages/jetpack-mu-wpcom we do need it. Unfortunately, since the
package went ESM-only, we have to jump through a hoop to be able to use
it from the CommonJS config file. Fortunately Webpack supports an async
function as the config.