Don't export wp_navigation posts by non-current users
This code doesn't work, and even if it did it's a hack that is probably
overly broad.

In the past a fallback navigation was created when one didn't exist.
This used to be created in `render_block_core_navigation` (see
WordPress/gutenberg#47684) but no longer happens.

This has lead to a number of blogs having this fallback navigation
created when an a12 visited their site. This means that the a12s details
- wpcom username, email address etc are in the generated export file.

It feels unhelpful to have this in the export file, partly because it
might need explaining if the legal time share an export as part of a
legal request, and partly because a12s data is shared to anyone
exporting an affected blog.

This change tries to remove the information when generating the export
file, but is somewhat hampered by a lack of relevent hooks and also by a
way of identifying current and former a12s. Instead it chooses a
relevant SQL generation, hackily identifies whether the query is the
relevant part of the export, and then modifies it to add some SQL
excluding wp_navigation post types that were created by users that are
not currently part of the blog.

See also https://linear.app/a8c/issue/DOTCOM-13020