ci: Fix conditions for skipping the coverage upload
There are two repository checks we need:

* We only want uploads from the actual monorepo, not any fork, so check
  for `github.repository === 'Automattic/jetpack'`.
* We also can't do the upload for PRs from forks, even when they target
  the monorepo, because it requires access to secrets. That check is
  `github.event.pull_request.head.repo.full_name === github.event.pull_request.base.repo.full_name`

We were missing the second one. Either we overlooked it entirely, or we
were confused into thinking the first check covered both needs.

Also, for good measure, tighten the condition for
`github.ref == 'refs/heads/trunk'` to only apply on pushes. It shouldn't
be possible to get a pull_request event for that, but it doesn't hurt to
check.