tools, ci: Fix tests on plugin release branches
Plugin tests on release branches for plugins that use
packages/test-environment have been failing since #41515 because in the
release branch we now remove the non-mirrored package.

To fix this, let's record removed packages when creating the release
branch, and restore them to composer.json before trying to run the
tests.