Blocks: In scan_delimiters extract static method
Ports 185482-ghe-Automattic/wpcom

When `scan_delimiters()` is attempting to find the end of an HTML
comment, it’s calling a convenience method which is also a closure
in the code (or an anonymous function). This carries a marginal
performance overhead and limits reuse.

In this change that method is extracted onto the class as its own static
method; this should simplify reasoning about the method itself and
provide a perofmrnace lift to the calling code.