CRM: Update FullCalendar from v3 to v6 (#43970)
* Fix zeroBSCRM_getLocale

If one ever requested the short version of the locale, it would always return the short version on subsequent calls, even if requesting the full version. We now cache the full version, and derive the short version from that as needed.

* Bump to FullCalendar 5

This also temporarily removes avatar and completion checkmark event UI modifications, since styling has changed significantly between versions and will likely continue to do so as we continue upgrades.

* Bump to FullCalendar 6

* Remove FullCalendar style overrides

* Add styles and task completion

* Add changelog

* Remove peer dependency hack

* Update pnpm-lock.yaml

* Update Phan baseline

* Reduce explicit dependencies

* Update pnpm-lock.yaml

* Prevent error on task edit page

* Fix links on list view

* Adjust header styles for small screens

* More adjustments

* Show user dropdown for all task-perms people