Blocks: Add Block_Scanner class, replacing Block_Delimiter
The `Block_Delimiter` class introduced `next_delimiter()` and
`scan_delimiters()`, which made it possible to parse the block structure in a document in a memory-efficient way. Unfortunately, fundamental
choices for the interface, namely returning a new class instance on
every block delimiter and relying on a generator function, limited the CPU performance fronteir of that class as a replacement for `parse_blocks()`.

This new class introduces `Block_Scanner`, more directly-modeled after the HTML API and informed by refactors incorporating `Block_Delimiter`.
This class mutates itself and requires a new instance before scanning.
The tradeoff is that it’s much faster running while maintaining the same near-zero memory overhead.

A new class is introduced due to the scale of change in the interface
and in order to provide seamless refactoring of code already relying
on `scan_delimiters()`.