fix: Ensure editor assets endpoint excludes disallowed plugin assets (#44093)
* fix: Ensure editor assets endpoint excludes disallowed plugin assets

The previous tests were false-positives, as the `disallowed-script`
registration had no impact on the environment and test outcome.

After fixing the test setup, it was evident the de-registration of
assets was invoked too earlier in the logic.

* feat: Allow in-progress Jetpack plugin assets

The `jetpack-dev` plugin is used to load in-progress work. Allowing its
assets ensures one can load assets from this helpful plugin.

* changelog