Replace remaining uses of lodash in js-packages/charts
All that was left was `merge` in two files, which we can replace with
the smaller `deepmerge` package.

Note `deepmerge` merges arrays differently by default (deepmerge
concatenates while lodash merges values at the same index), but as far
as I can tell that makes no difference to the existing uses here.

This saves about 68K uncompressed / 24K compressed if bundled alone, or
saves having a dependency on the entire lodash bundle if
dependency-extracted for WordPress.