fix: Reinstate Jetpack assets to editor assets endpoint
Recent changes fixed de-registering of assets not present on the allow
list. However, it was overlooked that certain Jetpack assets were then
erroneously excluded.

Namely, the `jetpack-blocks-editor` assets were excluded as their
`jetpack-blocks-assets-base-url` dependency was also excluded. The
latter was excluded as it is an "inline" script without a `src`
attribute. The lack of a `src` meant the path-based exclusion
erroneously de-registered the script.

Using a handle-based exclusion means that both inline and external
scripts are correct assessed by the plugin filtering logic.