tools: Update json files atomically in project-version.sh (#44273)
We had some builds fail because one process was trying to read a
`package.json` file while `tools/project-version.sh` in another process
was in the middle of `echo "$JSON" > "$FILE"`. Avoid this by having
`tools/project-version.sh` write to a temporary filename and then `mv`
it over the original, which should be an atomic filesystem operation.