Cleanup lodash `get` function (#44269)
The `get( obj, array )` style is straightforward, just use optional
chaining for all the elements in the array.

The `get( obj, string )` needs a little more care, as it will parse the
string: `get( obj, 'foo.bar[0].baz' )` → `obj?.foo?.bar?.[0]?.baz`. If
the string is or interpolates a variable that can be a little risky,
although it looks like everywhere where that has been done in existing
code isn't expecting to be getting a path.

Also this cleans up the one use of `result`, which is basically `get`
with a little extra magic we don't seem to make use of.